chore: improve CircularProgress, ClickOutside, Code exports
